change byte array to not be associated with a type
