+++ fastcrypto/src/lib.rs
@@ -89,7 +89,6 @@ pub mod secp256k1;
 pub mod secp256r1;
 
 // Other tooling
-pub mod bytes_representation;
 pub mod encoding;
 pub mod error;
 pub mod private_seed;
