+++ fastcrypto/src/serde_helpers.rs
@@ -3,6 +3,7 @@
 
 use base64ct::Encoding as _;
 use serde::{
+    de,
     de::{Deserializer, Error},
     ser::Serializer,
     Deserialize, Serialize,
@@ -120,16 +121,24 @@ impl<'de> DeserializeAs<'de, ed25519_consensus::Signature> for Ed25519Signature 
     }
 }
 
-//////////////////////////////////////////////////////////////////////////////////////////////////
-// Serialization of objects as bytes.
-// - for is_human_readable() serializers it returns base64 encoded bytes.
-// - else, it returns the bytes as a fixed size array.
-//
-// Objects that do not store a cached version of the serialized object should implement
-// [ToFromByteArray] and call [serialize_deserialize_with_to_from_byte_array].
-// Objects that do store a cached version should implement [ToFromBytes] and call
-// [serialize_deserialize_with_to_from_bytes].
-//
+///
+/// Serialization of internal types.
+///
+/// Every type's serialization should work as following:
+/// - Return base64 encoded bytes for is_human_readable() serializers.
+/// - Else, return the bytes as a fixed size array. This should be the canonical representation of
+///   the object (e.g., as defined in RFCs).
+///
+/// Types that do store a cached version of the serialized object should implement [ToFromBytes]
+/// and call [serialize_deserialize_with_to_from_bytes].
+///
+/// Types that do not store a cached version of the serialized object should implement
+/// [ToFromByteArray] and call [serialize_deserialize_with_to_from_byte_array].
+///
+/// Note that in theory internal types should not be exposed via APIs and thus never be serialized
+/// with is_human_readable(). Instead, external types should be used (see
+/// [generate_bytes_representation]).
+///
 
 // Serde treats arrays larger than 32 as variable length arrays, and adds the length as a prefix.
 // Since we want a fixed size representation, we wrap it in this helper struct and use serde_as.
@@ -215,4 +224,135 @@ macro_rules! serialize_deserialize_with_to_from_bytes {
             }
         }
     };
+}
+
+///
+/// External types.
+///
+/// [BytesRepresentation] is a basic wrapper for storing bincode serialized objects.
+///
+/// To be used in external interfaces instead of the internal types.
+/// - Derive by calling [generate_bytes_representation].
+/// - Uses Base64 when serialized with a human readable serializer, and raw bytes otherwise.
+///
+/// Note that in theory external types should not be stored and only be serialized with
+/// is_human_readable(). For storage and usage, internal types should be used (see
+/// above).
+///
+
+#[derive(Clone, Debug, PartialEq, Eq)]
+pub struct BytesRepresentation<const N: usize>(pub [u8; N]);
+
+/// Macro for generating a new alias for BytesRepresentation with the given $length, and From
+/// functions for both directions.
+/// - $type - the source type.
+/// - $length - byte length of the source type when serialized using bincode.
+/// - $new_type - the alias name.
+///
+// TODO: Can we deduce $new_type from $type?
+#[macro_export]
+macro_rules! generate_bytes_representation {
+    ($type:ty, $length:ident, $new_type:ident) => {
+
+        pub type $new_type = BytesRepresentation<$length>;
+
+        impl TryFrom<&BytesRepresentation<$length>> for $type {
+            type Error = FastCryptoError;
+
+            fn try_from(value: &BytesRepresentation<$length>) -> Result<Self, Self::Error> {
+                let o: $type =
+                    bincode::deserialize(&value.0).map_err(|_| FastCryptoError::InvalidInput)?;
+                Ok(o)
+            }
+        }
+
+        impl From<&$type> for BytesRepresentation<$length> {
+            fn from(value: &$type) -> Self {
+                let buffer = bincode::serialize(value).unwrap();
+                let buffer_len = buffer.len();
+                Self(
+                    // The following error should happen only in case the object was defined with
+                    // the wrong length, thus cannot be handled in runtime.
+                    buffer.try_into().expect(
+                        format!(
+                            "BytesRepresentation of length {} defined with invalid serialized length {}",
+                            $length, buffer_len
+                        )
+                        .as_ref(),
+                    ),
+                )
+            }
+        }
+    };
+}
+
+impl<const N: usize> Serialize for BytesRepresentation<N> {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: serde::Serializer,
+    {
+        match serializer.is_human_readable() {
+            true => Base64::encode(self.0).serialize(serializer),
+            false => SerializationHelper::<N>(self.0).serialize(serializer),
+        }
+    }
+}
+
+impl<'de, const N: usize> Deserialize<'de> for BytesRepresentation<N> {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: de::Deserializer<'de>,
+    {
+        let bytes: [u8; N] = match deserializer.is_human_readable() {
+            true => {
+                let s = String::deserialize(deserializer)?;
+                let decoded =
+                    Base64::decode(&s).map_err(|_| de::Error::custom("Base64 decoding failed"))?;
+                if decoded.len() != N {
+                    return Err(de::Error::custom(format!(
+                        "Invalid buffer length {}, expecting {}",
+                        decoded.len(),
+                        N
+                    )));
+                }
+                decoded.try_into().unwrap()
+            }
+            false => {
+                let helper: SerializationHelper<N> = Deserialize::deserialize(deserializer)?;
+                helper.0
+            }
+        };
+        Ok(Self(bytes))
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::groups::bls12381::{G1Element, G1ElementAsBytes, G_1_ELEMENT_BYTE_LENGTH};
+    use crate::groups::GroupElement;
+
+    #[derive(Serialize, Deserialize)]
+    struct Dummy<T> {
+        key: T,
+    }
+
+    #[test]
+    fn test_serializations() {
+        let g1 = G1Element::generator();
+        let b64 = G1ElementAsBytes::from(&g1);
+        let d1 = Dummy::<G1ElementAsBytes> { key: b64 };
+        // Test that we are not adding extra fields.
+        assert_eq!(
+            serde_json::to_string(&d1).unwrap(),
+            r#"{"key":"l/HTpzGX15QmlWOMT6msD8NojE+XdLkFoU46PxcbrFhsVeg/+Xoa7/s68ArbIsa7"}"#
+        );
+        // Test that we don't add extra bytes on top of the actual serialized data.
+        let ser = bincode::serialize(&d1).unwrap();
+        assert_eq!(G_1_ELEMENT_BYTE_LENGTH, ser.len());
+        // Check we can go back correctly.
+        let d2: Dummy<G1ElementAsBytes> = bincode::deserialize(&ser).unwrap();
+        let g2 = G1Element::try_from(&d2.key).unwrap();
+        assert_eq!(g1, g2);
+    }
 }
